Move StudentRepository to repository package
